seperated user created nodes and vipp nodes
