+++ ryven-editor/ryven/vipp_nodes/__init__.py
@@ -1,0 +1,3 @@
+from .nodes import *  # re-export nodes if this package is loaded directly
+
+
